Split FeatureContext into two ApiContext and WebContext
